requirements.txt
asgiref==3.6.0
Django==4.1
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
